% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_icils_iea.R
\name{download_icils_iea}
\alias{download_icils_iea}
\title{Download and import an empty ICILS public use file.}
\usage{
download_icils_iea(
  year = c("2018", "2013"),
  data_type = c("stud_dat", "stud_nat_dat", "stud_int_dat", "teach_dat", "teach_nat_dat",
    "teach_int_dat", "school_dat", "school_nat_dat", "school_int_dat", "it_dat")
)
}
\arguments{
\item{year}{Year of the assessment.}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty ICILS public use file (containing only the first data row) from the \href{https://www.iea.nl/data-tools/repository/icils}{IEA homepage}.
}
\details{
The function downloads a zip file from the IEA homepage into a temporary directory,
unzips it and imports the data with only a single data row via \code{\link[haven]{read_sav}}.
For downloading full ICILS data sets see the \href{https://cran.r-project.org/package=EdSurvey}{EdSurvey} package.
The data is imported as a \code{GADSdat} object.
}
\examples{
iea_icils <- download_icils_iea(year = "2018",
                        data_type = "stud_dat")
}
