% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_iglu_iea.R
\name{download_iglu_iea}
\alias{download_iglu_iea}
\title{Download and import an empty IGLU public use file.}
\usage{
download_iglu_iea(
  year = c("2016", "2011", "2006", "2001"),
  data_type = c("stud_par_dat", "teach_ger_dat", "teach_math_dat", "teach_gen_dat",
    "teach_dat", "teach_stud_dat", "school_dat", "tracking", "testscores")
)
}
\arguments{
\item{year}{Year of the assessment.}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty IGLU public use file (containing only the first data row) from the \href{https://www.iea.nl/data-tools/repository/pirls}{IEA homepage}.
}
\details{
The function downloads a zip file from the IEA homepage into a temporary directory,
unzips it and imports the data with only a single data row via \code{\link[haven]{read_sav}}.
For downloading full IGLU data sets see the \href{https://cran.r-project.org/package=EdSurvey}{EdSurvey} package.
The data is imported as a \code{GADSdat} object.
}
\examples{
iea_iglu <- download_iglu_iea(year = "2016",
                        data_type = "stud_par_dat")
}
