% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pisa_oecd.R
\name{download_pisa_oecd}
\alias{download_pisa_oecd}
\title{Download and import an empty PISA public use file.}
\usage{
download_pisa_oecd(
  year = c("2018", "2015", "2012", "2009", "2006", "2003", "2000"),
  data_type = c("stud_dat_9kl", "stud_dat_15j", "stud_par_dat", "stud_par_dat_9kl",
    "stud_par_dat_15j", "par_dat", "par_dat_9kl", "par_dat_15j", "teach_dat",
    "teach_dat_9kl", "teach_dat_15j", "school_dat", "school_dat_9kl", "school_dat_15j",
    "matching", "timing")
)
}
\arguments{
\item{year}{Year of the assessment.}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty PISA public use file (zero rows, but full metadata)
from the \href{https://www.oecd.org/en/data/datasets/}{OECD homepage}.
}
\details{
For PISA 2000–2012 the function uses the \pkg{EdSurvey} workflow (\code{downloadPISA},
\code{readPISA}) and converts the resulting file formats into a \code{GADSdat} object
with zero rows but full metadata.
For PISA 2015 and 2018 the function downloads a zip file from the OECD homepage into a
temporary directory, unzips it and imports the data via
\code{\link[haven]{read_sav}} or \code{\link[utils]{read.table}}.

For downloading full PISA data sets see the \href{https://cran.r-project.org/package=EdSurvey}{EdSurvey} package.
}
\examples{
oecd_pisa <- download_pisa_oecd(year = "2018",
                        data_type = "school_dat")
}
