% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_timss_iea.R
\name{download_timss_iea}
\alias{download_timss_iea}
\title{Download and import an empty TIMSS public use file.}
\usage{
download_timss_iea(
  year = c("2019", "2015", "2011", "2007"),
  data_type = c("stud_par_dat", "teach_dat", "teach_stud_dat", "school_dat", "tracking")
)
}
\arguments{
\item{year}{Year of the assessment.}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty TIMSS public use file (containing only the first data row) from the \href{https://www.iea.nl/data-tools/repository/icils}{IEA homepage}.
}
\details{
The function downloads a zip file from the IEA homepage into a temporary directory,
unzips it and imports the data with only a single data row via \code{\link[haven]{read_sav}}.
For downloading full TIMSS data sets see the \href{https://cran.r-project.org/package=EdSurvey}{EdSurvey} package.
The data is imported as a \code{GADSdat} object.
}
\examples{
iea_timss <- download_timss_iea(year = "2019",
                        data_type = "stud_par_dat")
}
