% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tvd_oecd.R
\name{download_tvd_oecd}
\alias{download_tvd_oecd}
\title{Download and import an empty TVD public use file.}
\usage{
download_tvd_oecd(
  data_type = c("stud_dat", "teach_dat", "teach_log_dat", "video_timss_dat", "video_dat",
    "video_subj_dat", "video_teach_dat", "video_third_dat", "artefact")
)
}
\arguments{
\item{data_type}{Type of the data.}
}
\description{
Download and import an empty TVD public use file (containing only the first data row) from the \href{https://www.oecd.org/en/data/datasets/}{OECD homepage}.
}
\details{
The function downloads a zip file from the OECD homepage into a temporary directory,
unzips it and imports the data with only a single data row via \code{\link[utils]{read.csv}}.
For downloading full TVD data sets see the \href{https://cran.r-project.org/package=EdSurvey}{EdSurvey} package.
The data is imported as a \code{GADSdat} object.
}
\examples{
oecd_tvd <- download_tvd_oecd(data_type = "stud_dat")
}
